Fixed a bug in migrating aws-settings (#977)
The S3 setting `endpoint` is normally not set at all in a production environment.
But the migration to engine-settings for this value was setting it to an empty
string. This was resulting in S3 trying to point to the endpoint <empty string>,
which would fail on all commands.

Making engine-settings' value field be nullable/undefineable would cause some
typing concerns elsewhere in the codebase. Instead, tweaked the logic to take
advantage of empty string's falseyness so that the endpoint resolution will
be undefined if it's an empty string.

Resolves IR-6663 IR-6665 IR-6669 IR-6673